golangci-lint v2 setup
Add .golangci.yml with a curated set of linters (staticcheck, govet,
errcheck, revive, gosec, gocritic, bodyclose, errorlint, unparam,
unconvert, usestdlibvars, usetesting, formatters), wire it into the
Makefile (lint, lint-fix) and CI as a separate job, drop the standalone
staticcheck step. Fix the issues surfaced by the initial run: gofmt,
errors.Is for io.EOF comparisons, error-checked Auth.Apply, removed
unnecessary int32 conversions and a duplicated branch in BuildURL.

+++ client/cmd_sync.go
@@ -14,7 +14,7 @@ func (c *Client) Sync(ctx context.Context, user string, req *eas.SyncRequest) (*
 	if err := c.do(ctx, CmdSync, user, req, &resp); err != nil {
 		return nil, err
 	}
-	if resp.Status != 0 && resp.Status != int32(eas.SyncStatusSuccess) {
+	if resp.Status != 0 && resp.Status != eas.SyncStatusSuccess {
 		return &resp, &StatusError{Command: "Sync", Status: resp.Status}
 	}
 	return &resp, nil
